feat(scalars): add months view and year view for date-picker
