+++ packages/design-system/src/scalars/components/date-picker-field/date-picker-field.tsx
@@ -1,4 +1,4 @@
-import React, { forwardRef } from "react";
+import { forwardRef } from "react";
 import { FieldCommonProps } from "../types";
 import { DateFieldValue } from "./types";
 import { withFieldValidation } from "../fragments/with-field-validation";
@@ -8,9 +8,8 @@ import { FormGroup } from "../fragments/form-group";
 import { FormLabel } from "../fragments/form-label";
 import { FormMessageList } from "../fragments/form-message";
 import { FormDescription } from "../fragments/form-description";
-import { Calendar } from "../fragments/calendar/calendar";
+import { Calendar } from "./subcomponents/calendar/calendar";
 import { cn } from "@/scalars/lib/utils";
-import { format } from "date-fns";
 import { useDatePickerField } from "./use-date-picker-field";
 
 export interface DatePickerFieldProps extends FieldCommonProps<DateFieldValue> {
