+++ packages/design-system/src/scalars/components/date-picker-field/subcomponents/utils.ts
@@ -1,0 +1,14 @@
+export const MONTHS = [
+  "January",
+  "February",
+  "March",
+  "April",
+  "May",
+  "June",
+  "July",
+  "August",
+  "September",
+  "October",
+  "November",
+  "December",
+];
