+++ packages/design-system/src/scalars/components/date-picker-field/types.ts
@@ -1,1 +1,2 @@
 export type DateFieldValue = string | Date | null;
+export type DatePickerView = "years" | "months" | "days";
